refactor: split GenreGuide.tsx into smaller modules
Split the 1,041-line GenreGuide component into 13 focused modules:

- types.ts (31) — GenreGuideProps, SortKey, EnrichedLens, SCORED_GENRES
- helpers.ts (26) — sceneLabel, evHeaderLabel, fmtIso, sortIndicator
- useGenreState.ts (160) — all state, handlers, derived values
- useEnrichedLenses.ts (102) — filter + enrich + sort logic
- GenreControls.tsx (119) — mount/FL/ISO/ND/magnification chips
- FilterSelect.tsx (31) — reusable filter dropdown component
- GenreFilterPanel.tsx (67) — genre-conditional filter panel
- genreColumns.ts (116) — per-genre column definitions
- GenreTable.tsx (59) — desktop table with sortable headers
- GenreRowCells.tsx (109) — per-genre table row cells
- GenreCards.tsx (51) — mobile card layout
- GenreFooter.tsx (28) — scoring methodology text
- GenreGuide.tsx (109) — thin orchestrator

All existing tests pass unchanged.

Closes #288
